backend/requirements.txt
# fastapi
# uvicorn[standard]
# pydantic
# pydantic-settings
# python-multipart
# pymongo
# motor
# ffmpeg-python
# opencv-python-headless
# librosa
# soundfile
# numpy
# pandas
# sentence-transformers==2.2.2
# huggingface_hub==0.13.4
# openai-whisper
# transformers
# tqdm
# mediapipe==0.10.14
# pytesseract
# Pillow
# scikit-learn
# scipy
# python-dotenv
# aiofiles
# redis
# celery

fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
pymongo
motor
ffmpeg-python
opencv-python-headless
librosa
soundfile
numpy
pandas
sentence-transformers==2.2.2
huggingface_hub==0.13.4
openai-whisper
transformers
tqdm
mediapipe==0.10.14
pytesseract
Pillow
scikit-learn
scipy
python-dotenv
aiofiles
redis
celery
google-generativeai